Revert nixos-rebuild using flakes
I'll look at nix-env -> nix profile, but I'm not seeing the benefit at
the nixos-rebuild level.

It's annoying to me that nixos-rebuild insists on trying to
update/commit lock files in a git repo, then messes up the chown since
it's running as root. This can be worked around (see the flags passed to
autoupgrade, or maybe the --use-remote-sudo hack), but when not using a
flake.lock, channels do a better job of tracking versioning.

If I went all-in on flake.lock, there'd be no obvious way to run
autoupgrade except with the flags reverted in this commit. If I do that,
I will accidentally downgrade my system if I run an nixos-rebuild
manually without updating flake.lock. Alternatively, I could not do an
automatic flake equivalent of --upgrade, but I like that functionality.

I could run a systemd unit with user privileges that periodically
updates flake.lock, then commits and maybe pushes the change, but I'd
rather not have automated git changes in my config for now.

n.b. Had to boot into old system build to rerun nixos-rebuild without
flakes.